--- secret-service/pkg/backend/scopebackend.go
@@ -1,5 +1,0 @@
-package backend
-
-type ScopeBackend interface {
-	GetScopes() ([]string, error)
-}
