+++ secret-service/pkg/backend/secretbackend.go
@@ -6,12 +6,19 @@ import (
 
 const DefaultNamespace = "keptn"
 
-//go:generate moq -pkg fake -out ./fake/secretbackend_mock.go . SecretBackend
-type SecretBackend interface {
+type SecretManager interface {
 	CreateSecret(model.Secret) error
 	UpdateSecret(model.Secret) error
 	DeleteSecret(model.Secret) error
 	GetSecrets() ([]model.GetSecretResponseItem, error)
+}
 
-	ScopeBackend
+type ScopeManager interface {
+	GetScopes() ([]string, error)
+}
+
+//go:generate moq -pkg fake -out ./fake/secretbackend_mock.go . SecretBackend
+type SecretBackend interface {
+	SecretManager
+	ScopeManager
 }
